net: shell: Common header for all shell commands
Some useful APIs and macros for all net shell commands
to use.
